+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 ### Added
  - JAMES-3524 Support symmetric encryption support on top of BlobStore
  
+### Removed
+ - JAMES-3578 Drop Cassandra schema version prior version 8 (see upgrade instructions)
+ 
 ### Third party software
  - Upgrading to Apache Tika 1.26 is recommended
      - 1.25 and before are subject to CVE-2021-28657 CVE-2021-27906 CVE-2021-27807